% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_dynamic_transacted}
\alias{collect_dynamic_transacted}
\title{FUNCTION: collect_dynamic_transacted}
\usage{
collect_dynamic_transacted(data)
}
\description{
DESCRIPTION: Collect only the data values that are for the "dynamic" or "transacted" removers
}

