% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_fast_databases}
\alias{collect_fast_databases}
\title{FUNCTION: collect_fast_databases}
\usage{
collect_fast_databases(data)
}
\description{
DESCRIPTION: Collect only the two fast databases (i.e., HyperSQL and SQLite)
}

