% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_hypersql_postgres}
\alias{collect_hypersql_postgres}
\title{FUNCTION: collect_hypersql_postgres}
\usage{
collect_hypersql_postgres(d)
}
\description{
DESCRIPTION: Collect the data values that are related to two DBMSs, both Postgres and HyperSQL
}

