% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_minus_stillborn_equivalent_redundant}
\alias{collect_minus_stillborn_equivalent_redundant}
\title{FUNCTION: collect_minus_stillborn_equivalent_redundant}
\usage{
collect_minus_stillborn_equivalent_redundant(d)
}
\description{
DESCRIPTION: Collect only the data for -(S), -(S+I+E), and -(S+I+E+R). This is used to remove the
-(S+I) data from the strict dbms data because it is identical to the -(S) data.
}

