% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_mutants_equivalent}
\alias{collect_mutants_equivalent}
\title{FUNCTION:  collect_mutants_equivalent}
\usage{
collect_mutants_equivalent(d)
}
\description{
DESCRIPTION: Collect the mutants that are run and yet are equivalent
}

