% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_mutants_impaired}
\alias{collect_mutants_impaired}
\title{FUNCTION:  collect_mutants_impaired}
\usage{
collect_mutants_impaired(d)
}
\description{
DESCRIPTION: Collect the mutants that are run and yet are impaired
}

