% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_mutants_killed_count}
\alias{collect_mutants_killed_count}
\title{FUNCTION: collect_mutants_killed_count}
\usage{
collect_mutants_killed_count(d)
}
\description{
DESCRIPTION: Collect the mutants that are run and are killed
}

