% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_mutants_redundant}
\alias{collect_mutants_redundant}
\title{FUNCTION: collect_mutants_redundant}
\usage{
collect_mutants_redundant(d)
}
\description{
DESCRIPTION: Collect the mutants that are run and yet are redundant
}

