% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_mutants_stillborn}
\alias{collect_mutants_stillborn}
\title{FUNCTION:  collect_mutants_stillborn}
\usage{
collect_mutants_stillborn(d)
}
\description{
DESCRIPTION: Collect the mutants that are stillborn
}

