% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_remove_mutants_impaired_equivalent}
\alias{collect_remove_mutants_impaired_equivalent}
\title{FUNCTION:  collect_remove_mutants_impaired_equivalent}
\usage{
collect_remove_mutants_impaired_equivalent(d)
}
\description{
DESCRIPTION: Collect the mutants that are run and yet are equivalent, includes imparied mutants and remove them
}

