% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_remove_mutants_impaired_equivalent_redundant}
\alias{collect_remove_mutants_impaired_equivalent_redundant}
\title{FUNCTION:  collect_remove_mutants_impaired_equivalent_redundant}
\usage{
collect_remove_mutants_impaired_equivalent_redundant(d)
}
\description{
DESCRIPTION: Collect the mutants that are run and yet are redundant, includes imparied and equivalent mutants and remove them
}

