% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_removers_dynamic_transacted}
\alias{collect_removers_dynamic_transacted}
\title{FUNCTION: collect_removers_dynamic_transacted}
\usage{
collect_removers_dynamic_transacted(data)
}
\description{
DESCRIPTION: Collect only the data values that are for "remover" aspects of the mutation pipeline
}

