% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_removers_equivalent}
\alias{collect_removers_equivalent}
\title{FUNCTION: collect_removers_equivalent}
\usage{
collect_removers_equivalent(d)
}
\description{
DESCRIPTION: Collect the removers that are for equivalent mutants
We need to first remove stillborn and impaired because mutant removal is an iterative process.
The removal processes are not independent of eachother --- they build on the last stage.
}

