% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_removers_impaired}
\alias{collect_removers_impaired}
\title{FUNCTION: collect_removers_impaired}
\usage{
collect_removers_impaired(d)
}
\description{
DESCRIPTION: Collect the removers that are for impaired mutants. Only SQLite dbms has impaired mutants
We need to first remove stillborn because mutant removal is an iterative process.
The removal processes are not independent of eachother --- they build on the last stage.
}

