% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_removers_redundant}
\alias{collect_removers_redundant}
\title{FUNCTION: collect_removers_redundant}
\usage{
collect_removers_redundant(d)
}
\description{
DESCRIPTION: Collect the removers that are for redundant mutants
We need to first remove stillborn, impaired and equivalent because mutant removal is an iterative process.
The removal processes are not independent of eachother --- they build on the last stage.
}

