% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_removers_static}
\alias{collect_removers_static}
\title{FUNCTION: collect_removers_static}
\usage{
collect_removers_static(data)
}
\description{
DESCRIPTION: Collect only the data values that are for "remover" aspects of the mutation pipeline; could be refactored!
}

