% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_removers_static_hypersql}
\alias{collect_removers_static_hypersql}
\title{FUNCTION:  collect_removers_static_hypersql}
\usage{
collect_removers_static_hypersql(d)
}
\description{
DESCRIPTION: Collect only the data values that are for static analysis with the HyperSQL DBMS
}

