% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_removers_static_postgres}
\alias{collect_removers_static_postgres}
\title{FUNCTION:  collect_removers_static_postgres}
\usage{
collect_removers_static_postgres(d)
}
\description{
DESCRIPTION: Collect only the data values that are for static analysis with the PostgreSQL DBMS
}

