% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_removers_stillborn}
\alias{collect_removers_stillborn}
\title{FUNCTION: collect_removers_stillborn}
\usage{
collect_removers_stillborn(d)
}
\description{
DESCRIPTION: Collect the removers that are for stillborn mutants SQLite dbms has NO stillborn mutants.
}

