% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_slow_databases}
\alias{collect_slow_databases}
\title{FUNCTION:  collect_slow_databases}
\usage{
collect_slow_databases(data)
}
\description{
DESCRIPTION: Collect the slow database (i.e., Postgres)
}

