% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_strict_databases}
\alias{collect_strict_databases}
\title{FUNCTION:  collect_strict_databases}
\usage{
collect_strict_databases(d)
}
\description{
DESCRIPTION: Collect the strict databases (i.e., Postgres and HyperSQL)
}

