% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_mutation_score_killed_total_count}
\alias{join_mutation_score_killed_total_count}
\title{FUNCTION: join_mutation_score_killed_total_count}
\usage{
join_mutation_score_killed_total_count(k, tot)
}
\description{
DESCRIPTION: Join together two mutation analysis data tables
}

