% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_stillborn_impaired_equivalent_redundant}
\alias{join_stillborn_impaired_equivalent_redundant}
\title{FUNCTION: join_stillborn_impaired_equivalent_redundant}
\usage{
join_stillborn_impaired_equivalent_redundant(k, i, e, r)
}
\description{
DESCRIPTION: Join together four data tables when matching on the schema and the dbms
}

