% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_analyse_pipeline}
\alias{read_analyse_pipeline}
\title{FUNCTION: read_analyse_pipeline
READ THE DATA FOR THE MUTATION ANALYSIS PIPELINE}
\usage{
read_analyse_pipeline()
}
\description{
DESCRIPTION: read in the data file containing the data about:
-- time taken to detect stillborn mutants using the
      -- Dynamic
      -- Transaction
      -- Static
methods. This data is recorded for all schemas and for all three DBMSs.
}

