% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_hypersql_avmdefaults}
\alias{read_hypersql_avmdefaults}
\title{FUNCTION: read_hypersql_avmdefaults
READ THE DATA FOR THE TIME TAKEN TO PERFORM MUTATION ANALYSIS}
\usage{
read_hypersql_avmdefaults()
}
\description{
DESCRIPTION: Read in the data file containing the data about:
-- time taken to run the mutation analysis for each
      -- Operator
      -- Type of mutant
}

