% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_postgres_avmdefaults}
\alias{read_postgres_avmdefaults}
\title{FUNCTION: read_postgres_avmdefaults}
\usage{
read_postgres_avmdefaults()
}
\description{
DESCRIPTION: Read in the data file containing the data about:
-- time taken to run the mutation analysis for each
      -- Operator
      -- Type of mutant
}

