% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_postgres_random}
\alias{read_postgres_random}
\title{FUNCTION: read_postgres_random}
\usage{
read_postgres_random()
}
\description{
DESCRIPTION: Read in the data file containing the data about:
-- time taken to run the mutation analysis for each
      -- Operator
      -- Type of mutant
}

