% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_sqlite_ineffective}
\alias{read_sqlite_ineffective}
\title{FUNCTION: read_sqlite_ineffective
READ THE DATA FOR THE INEFFECTIVE MUTANTS}
\usage{
read_sqlite_ineffective()
}
\description{
DESCRIPTION: read in the data file containing the data about:
-- time taken to detect stillborn mutants using the
      -- Dynamic
      -- Transaction
      -- Static
methods. This data is recorded for all schemas and for just SQLite DBMS.
}

