% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_strict_avmdefaults}
\alias{read_strict_avmdefaults}
\title{FUNCTION: read_strict_avmdefaults}
\usage{
read_strict_avmdefaults()
}
\description{
DESCRIPTION: Read in the two data files for the DBMS that are strict in their interpretation of the SQL
standard (e.g., HyperSQL and Postgres) since they are always grouped in the analysis of times and scores
}

