% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_mean_impaired_time}
\alias{summarise_mean_impaired_time}
\title{FUNCTION: summarise_mean_impaired_time}
\usage{
summarise_mean_impaired_time(d)
}
\description{
DESCRIPTION: Summarise the data for running the stillborn, impaired mutants for each schema
}

