% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_mutation_analysis_impaired_operator}
\alias{summarise_mutation_analysis_impaired_operator}
\title{FUNCTION: summarise_mutation_analysis_impaired_operator}
\usage{
summarise_mutation_analysis_impaired_operator(d)
}
\description{
DESCRIPTION: Summarise the time required for analysing impaired mutants during mutation analysis
}

