% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_mutation_analysis_stillborn_equivalent_redundant_time}
\alias{summarise_mutation_analysis_stillborn_equivalent_redundant_time}
\title{FUNCTION: summarise_mutation_analysis_stillborn_equivalent_redundant_time}
\usage{
summarise_mutation_analysis_stillborn_equivalent_redundant_time(d)
}
\description{
DESCRIPTION: Summarise the data for running the stillborn, equivalent and redundant mutants for each schema
}

