% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_mutation_analysis_stillborn_operator}
\alias{summarise_mutation_analysis_stillborn_operator}
\title{FUNCTION: summarise_mutation_analysis_stillborn_operator}
\usage{
summarise_mutation_analysis_stillborn_operator(d)
}
\description{
DESCRIPTION: Summarise the time required for analysing stillborn mutants during mutation analysis
This should not work! Stillborn mutants are not run during mutation analysis!
}

