% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_total_impaired_time}
\alias{summarise_total_impaired_time}
\title{FUNCTION: summarise_total_impaired_time}
\usage{
summarise_total_impaired_time(d)

summarise_total_impaired_time(d)
}
\description{
DESCRIPTION: Summarise the data associated with running the impaired mutants alone

DESCRIPTION: Summarise the data associated with running the impaired mutants alone
}

