% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_total_mutation_score}
\alias{summarise_total_mutation_score}
\title{FUNCTION: summarise_total_mutation_score}
\usage{
summarise_total_mutation_score(d)
}
\description{
DESCRIPTION: Summarise the data associated with the mutation score, associated with running every type of mutant
}

