% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_total_mutation_time}
\alias{summarise_total_mutation_time}
\title{FUNCTION: summarise_total_mutation_time}
\usage{
summarise_total_mutation_time(d)
}
\description{
DESCRIPTION: Summarise the data associated with the mutation time, associated with running every type of mutant
}

