% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_add_all_time}
\alias{transform_add_all_time}
\title{FUNCTION: transform_add_all_time}
\usage{
transform_add_all_time(d)
}
\description{
DESCRIPTION: Add the times running the entire mutation analysis, no subtractions allow, no additions from
other analysis either (adding in static analysis time would bias the results).
This is the same as all stillborn because stillborn mutants have already been removed
}

