% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_add_equivalent_redundant_time}
\alias{transform_add_equivalent_redundant_time}
\title{FUNCTION: transform_add_equivalent_redundant_time}
\usage{
transform_add_equivalent_redundant_time(d)
}
\description{
DESCRIPTION: The All time is lessened by not running equivalent or redundant mutants, but we add on the cost of
running two operators in the mutation analysis pipeline, the primary keys and the equivalent mutant remover
}

