% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_add_impaired_time}
\alias{transform_add_impaired_time}
\title{FUNCTION: transform_add_impaired_time}
\usage{
transform_add_impaired_time(d)
}
\description{
DESCRIPTION: The All time is lessened by not running impaired mutants, but we add on the cost of
}

