% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_equivalent_mutation_score}
\alias{transform_equivalent_mutation_score}
\title{FUNCTION: transform_equivalent_mutation_score}
\usage{
transform_equivalent_mutation_score(d)
}
\description{
DESCRIPTION: The All mutation scores are further increased in accuracy by not running impaired or equivalent mutants,
}

