% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_impaired_mutation_score}
\alias{transform_impaired_mutation_score}
\title{FUNCTION: transform_impaired_mutation_score}
\usage{
transform_impaired_mutation_score(d)
}
\description{
DESCRIPTION: The All score is made more accurate by not running impaired mutants
}

