% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_move_characteristics}
\alias{transform_move_characteristics}
\title{FUNCTION: transform_move_characteristics}
\usage{
transform_move_characteristics(d)
}
\description{
DESCRIPTION: Move the characteristics of a data row to the font of a data frame
}

