% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_move_timings}
\alias{transform_move_timings}
\title{FUNCTION: transform_move_timings}
\usage{
transform_move_timings(d)
}
\description{
DESCRIPTION: Move any variable that starts with the word "time" to the front of the data frame
}

