% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_redundant_mutation_score}
\alias{transform_redundant_mutation_score}
\title{FUNCTION: transform_redundant_mutation_score}
\usage{
transform_redundant_mutation_score(d)
}
\description{
DESCRIPTION: The All mutation score is further increased by not running impaired, equivalent or redundant mutants
}

