% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_remove_operator}
\alias{transform_remove_operator}
\title{FUNCTION: transform_remove_operator}
\usage{
transform_remove_operator(d)
}
\description{
DESCRIPTION: Remove the operator column, normally before preparing to join data frames together
}

