% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_remove_pipeline}
\alias{transform_remove_pipeline}
\title{FUNCTION: transform_remove_pipeline}
\usage{
transform_remove_pipeline(d)
}
\description{
DESCRIPTION: Remove the pipeline column, normally before preparing to join data frames together
}

