% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_replace_NA_with_zero_i}
\alias{transform_replace_NA_with_zero_i}
\title{FUNCTION: transform_replace_NA_with_zero_i}
\usage{
transform_replace_NA_with_zero_i(d)
}
\description{
DESCRIPTION: Remove all of the NA values and replace them with a 0 value because no IMPAIRED mutants
}

