% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_replace_NA_with_zero_s}
\alias{transform_replace_NA_with_zero_s}
\title{FUNCTION: transform_replace_NA_with_zero_s}
\usage{
transform_replace_NA_with_zero_s(d)
}
\description{
DESCRIPTION: Remove all of the NA values and replace them with a 0 value because no STILLBORN mutants
}

