% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_standardise_schema}
\alias{transform_standardise_schema}
\title{FUNCTION: transform_standardise_schema}
\usage{
transform_standardise_schema(d)
}
\description{
DESCRIPTION: Make sure that we always use the word "schema" instead of "casestudy"
}

