% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_combined}
\alias{visualise_combined}
\title{FUNCTION: visualise_combined}
\usage{
visualise_combined(n, top, middle, bottom, h, w)
}
\description{
VISUALISE EXECUTION TIME
}
\details{
DESCRIPTION: Create a plot that combines all of the two graphics for the timetaken data
This function will be used for the "detailed" plots i.e., individual schema plots and the
summary plots.
}

