% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_mutation_score_strict_avmdefaults}
\alias{visualise_mutation_score_strict_avmdefaults}
\title{FUNCTION: visualise_mutation_score_strict_avmdefaults}
\usage{
visualise_mutation_score_strict_avmdefaults(d, pos)
}
\description{
DESCRIPTION: Plot and save a graph that shows the mutation SCORE for all of
the STRICT DBMSs and with all of the different REMOVERS (e.g., -S, -(S+I+E), -(S+I+E+R)) across all of the CASESTUDIES
}

