% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_mutation_timetaken_strict_random_summary}
\alias{visualise_mutation_timetaken_strict_random_summary}
\title{FUNCTION: visualise_mutation_timetaken_strict_random_summary}
\usage{
visualise_mutation_timetaken_strict_random_summary(d, pos)
}
\description{
DESCRIPTION: Plot and save a graph that shows the mutation TIMETAKEN for all of
the STRICT DBMSs and with all of the different REMOVERS (e.g., All, -E, -(E+R)) across all of the CASESTUDIES,
customized for a specific DATA GENERATOR that will be displayed in the plots title.
}

