% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_plot_mutation_time}
\alias{visualise_plot_mutation_time}
\title{FUNCTION: visualise_plot_mutation_time}
\usage{
visualise_plot_mutation_time(d, generator, pos = "bottom")
}
\description{
DESCRIPTION: Generally plot a box and whisker plot of the mutation time taken for different
configurations; multiple functions will ultimately call this one to produce a different graph
In order for these individual schema plots to look good, height needs to be set to 10 in the visualise_graphic_save
function.
}

