% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_plot_mutation_time_summary}
\alias{visualise_plot_mutation_time_summary}
\title{FUNCTION: visualise_plot_mutation_time_summary}
\usage{
visualise_plot_mutation_time_summary(d, generator, pos = "bottom")
}
\description{
DESCRIPTION: Generally plot a box and whisker plot of the mutation time taken for different
configurations; multiple functions will ultimately call this one to produce a different graph
This function aggregates all times across all schemas and removes outliers. Particularly,
this function is usual to show the decrease in mutation analysis time across mutant removal
stages.
}

