% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_plot_timetaken}
\alias{visualise_plot_timetaken}
\title{FUNCTION:  visualise_plot_timetaken}
\usage{
visualise_plot_timetaken(data)
}
\description{
DESCRIPTION: Generally plot different types of box and whisker plots for the timetaken in mutant removal
}

