% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_plot_timetaken_stack_pipeline}
\alias{visualise_plot_timetaken_stack_pipeline}
\title{FUNCTION:  visualise_plot_timetaken_stack_pipeline}
\usage{
visualise_plot_timetaken_stack_pipeline(data)
}
\description{
DESCRIPTION: Generally plot different types of box and whisker plots for the timetaken in mutant removal,
essentially just changing the way that the facet_grid is produced, compared to the visualise_plot_timetaken
}

