% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_timetaken_combined}
\alias{visualise_timetaken_combined}
\title{FUNCTION: visualise_timetaken_combined}
\usage{
visualise_timetaken_combined(top, bottom_left, bottom_right)
}
\description{
VISUALISE EXECUTION TIME
}
\details{
DESCRIPTION: Create a plot that combines all of the three graphics for the timetaken data of STILLBORN mutant removal
}

