% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_timetaken_dynamic_transacted_fast}
\alias{visualise_timetaken_dynamic_transacted_fast}
\title{FUNCTION: visualise_timetaken_dynamic_transacted_fast}
\usage{
visualise_timetaken_dynamic_transacted_fast(data)
}
\description{
VISUALISE EXECUTION TIME
}
\details{
DESCRIPTION: visualise_timetaken_dynamic_transacted_fast: Plot and save the box and whisker plot for the amount of TIMETAKEN to remove
STILLBORN mutants in the mutation analysis pipeline, across all of the FAST DBMSs and all CASESTUDIES, for DYNAMIC and TRANSACTED
}

